Grade english strings by chi2 character frequency
Get a chi2 (sum of squared differences) between expected english
relative character frequencies and observed character frequencies in a
string. Try to decode the crypted string with different keys and sort
the resulting strings according to the one that minimises chi2. Seems to
work pretty well after applying a simple preselection to get rid of
strings with an unreasonably high non word character fraction.

Could imagine doing a likelihood version as well, maybe treating the
number of occurances of each character as coming from a poisson
distribution of the appropriate mean.